% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftpdl.R
\name{ftpdl}
\alias{ftpdl}
\title{Download a File from FTP URL}
\usage{
ftpdl(
  burl = NULL,
  dlfile = NULL,
  outwd = getwd(),
  fusr = NULL,
  fpwd = NULL,
  foverwrite = FALSE,
  fredllim = 1L,
  fdlretry = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{burl}{(character) FTP URL that should be scanned.}

\item{dlfile}{(character) name of the file to be downloaded.}

\item{outwd}{(character; optional) directory where downloaded files should be
saved. (Default: location obtained from getwd())}

\item{fusr}{(character; optional) username used to access the FTP URL.}

\item{fpwd}{(character; optional) password for the FTP URL.#'}

\item{foverwrite}{(logical; optional) controls whether the local copy of a
file to be downloaded should be overwritten or not. TRUE - local copy will
be overwritten; FALSE - local copy will not be overwritten, the downloaded
file will be saved with the prefix "new_" instead. (Default: FALSE)}

\item{fredllim}{(numeric; optional) controls how many attempts should be made
to redownload a file upon download failure. This parameter is to be used in
conjunction with the parameter dlretry. If dlretry is set to FALSE, setting
redllim will have no effect. (Default: 1L)}

\item{fdlretry}{(logical; optional) controls whether a file download should
be retried if it fails. Set to TRUE to enable retries. This parameter is
usually used in conjunction with the parameter redllim. (Default: FALSE)}

\item{verbose}{(logical) controls whether console messages should be
displayed (e.g., messages such as "Downloading file: "). Set to FALSE to
enable silent mode. (Default: TRUE)}
}
\value{
a character string ("FAILURE"/"SUCCESS") indicating whether the
  download was (un)successful.
}
\description{
Downloads the file matching the given file name from the given FTP URL. Used
internally by repchkdl.
}
\examples{
\dontrun{
#Fetch some files with repchkdl::ftpreadlinks
#This URL is a test FTP server, and contains random files!
#This test server's password changes frequently
#Check the password at dlptest.com (and update) before proceeding!
test <- ftpreadlinks(rurl = "ftp://ftp.dlptest.com",
rusr = "dlpuser@dlptest.com", rpwd = "SzMf7rTE4pCrf9dV286GuNe4N")

#Download just the first file in test as an example
#Output will be saved wherever getwd() is pointing to.
test$dlstats[1] <- ftpdl(burl = "ftp://ftp.dlptest.com",
                   dlfile = mylinks$flinks[i], verbose = TRUE,
                   fusr = "dlpuser@dlptest.com",
                   fpwd = "SzMf7rTE4pCrf9dV286GuNe4N")
}

}
