% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftpreadlinks.R
\name{ftpreadlinks}
\alias{ftpreadlinks}
\title{Get set of file names and sizes from FTP URL}
\usage{
ftpreadlinks(
  rurl = NULL,
  rusr = NULL,
  rpwd = NULL,
  rsizelb = NULL,
  rsizeub = NULL,
  rregex = NULL,
  verbose = FALSE
)
}
\arguments{
\item{rurl}{(character) FTP URL that should be scanned.}

\item{rusr}{(character; optional) username used to access the FTP URL.}

\item{rpwd}{(character; optional) password for the FTP URL.}

\item{rsizelb}{(numeric; optional) minimum file size (in bytes). Only files
equal to or above this threshold will be downloaded.}

\item{rsizeub}{(numeric; optional) maximum file size (in bytes). Files equal
to or above this threshold will NOT be downloaded.}

\item{rregex}{(character; optional) regular expression to indicate what file
additions at the URL should be considered for download. The regex format
follows stringr conventions.}

\item{verbose}{(logical) controls whether console messages should be
displayed (e.g., messages such as "Downloading file: "). Set to FALSE to
enable silent mode. (Default: TRUE)}
}
\value{
a data.frame containing the names of the files and their sizes (in
  bytes) matching the optional constraints (e.g., minimum/maximum file size
  thresholds) available at the FTP URL for download.
}
\description{
Accesses a FTP URL and returns a data.frame of files and their sizes in
bytes. The set of files returned can be filtered by file size and regular
expressions. Used internally by repchkdl.
}
\examples{
\dontrun{
#This URL is a test FTP server, and contains random files!
#This test server's password changes frequently
#Check the password at dlptest.com (and update) before proceeding!
#If this example returns a zero length data.frame, rerun without a regex
#It is possible that this test server might not have any .txt files at
#the time of running this example!
test <- ftpreadlinks(rurl = "ftp://ftp.dlptest.com",
                     rusr = "dlpuser@dlptest.com",
                     rpwd = "SzMf7rTE4pCrf9dV286GuNe4N",
                     rregex = ".*.txt")
}

}
